+++ src/main/scala/scodec/codecs/VectorCodec.scala
@@ -7,7 +7,7 @@ private[codecs] final class VectorCodec[A](codec: Codec[A], limit: Option[Int] =
 
   def sizeBound = limit match {
     case None => SizeBound.unknown
-    case Some(lim) => codec.sizeBound * lim
+    case Some(lim) => codec.sizeBound * lim.toLong
   }
 
   def encode(vector: Vector[A]) = Encoder.encodeSeq(codec)(vector)
